luna_api/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlmodel==0.0.22
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
apscheduler==3.10.4
python-multipart